+++ rpc/src/errors/mod.rs
@@ -1,0 +1,4 @@
+mod parse_error;
+mod rpc_error;
+pub use parse_error::ParseError;
+pub use rpc_error::RpcError;
