+++ rpc/src/responses/mod.rs
@@ -1,7 +1,7 @@
 pub mod balance;
 pub mod block_ids_in_chain;
 mod bulk_array;
-use crate::types::errors::ParseError;
+use crate::errors::ParseError;
 pub use balance::BalanceResponse;
 pub use block_ids_in_chain::BlocksInChainResponse;
