+++ rpc/src/types/mod.rs
@@ -1,5 +1,4 @@
 mod block;
 mod chain;
-pub mod errors;
 pub use block::Block;
 pub use chain::Chain;
